refactor: move sections dir to top (avoid indivisual page generation)
