+++ pages/template.vue
@@ -5,8 +5,7 @@
 </template>
 
 <script>
-// import TextCenterST from '@/components/templates/TextCenterST/TextCenterST.vue'
-import TemplateShowcaseSec from '@/pages/sections/template/TemplateShowcaseSec/TemplateShowcaseSec/'
+import TemplateShowcaseSec from '@/sections/template/TemplateShowcaseSec/TemplateShowcaseSec/'
 export default {
   name: 'TemplatePage',
   components: { TemplateShowcaseSec },
