+++ sections/KeyVisualSec/script.js
