+++ sections/KeyVisualSec/style.scss
